minor changes and cleaning
added some buttons to the ContentUnavailableView in the ShoppingListView to add a new item to the shopping list directly, or to navigate over to the AllMyItemsView to pick items there.  this works on the iPad in NavigationSplitView, but does not (yet) work in compact tab views.

Location.append(:Item) now renamed to Location.addToItems(:Item).  a reciprocal Location.removeFromItems(:Item) has been added, and the relationship Location.itemsOptional is now marked fileprivate.

ReadMe updated.